requirements.txt
pytest==7.2.1
python-decouple==3.7
requests==2.28.2
